initial implementation of partial expansion of json -> hierlist
When loading large json files building the hierlist takes a extremely
long time, locking the interface. This commmit modifies how the hlist
is built so that nested json objects are only loaded as hierlists when
they are opened. This significantly improves performance.

One optimization that has not beend implemented is to remove rendered
objects from the unexpanded cache once they have been rendered.